progress on Annotation type consolidation
