+++ apps/frontend/src/components/annotation-popups/ResolvedReferencePopup.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { useRouter } from 'next/navigation';
 import { PopupContainer, PopupHeader, SelectedTextDisplay, EntityTypeBadges } from './SharedPopupElements';
 import { buttonStyles } from '@/lib/button-styles';
-import type { ReferenceAnnotation, AnnotationUpdate, TextSelection } from '@/types/annotation';
+import type { ReferenceAnnotation, AnnotationUpdate, TextSelection } from '@semiont/core-types';
 
 interface ResolvedReferencePopupProps {
   isOpen: boolean;
