+++ apps/frontend/src/components/annotation-popups/__tests__/StubReferencePopup.test.tsx
@@ -4,7 +4,7 @@ import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { StubReferencePopup } from '../StubReferencePopup';
 import { useRouter } from 'next/navigation';
-import type { ReferenceAnnotation, TextSelection } from '@/types/annotation';
+import type { ReferenceAnnotation, TextSelection } from '@semiont/core-types';
 
 // Mock next/navigation
 vi.mock('next/navigation', () => ({
